% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguindown.R
\docType{package}
\name{thesisdown}
\alias{penguindown}
\alias{penguindown-package}
\title{penguindown: A package for creating undergraduate, Masters, and PhD theses
using R Markdown}
\description{
Using the thesisdown package and LaTeX and Word thesis templates from Youngstown State University, this package aims to encourage reproducible research using simple Markdown syntax while embedding all of the R code to produce plots and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: G. Jay Kerns \email{gkerns@ysu.edu}

}
